backend/requirements.txt
azure-functions
agent-framework --pre
agent-framework-azurefunctions --pre
azure-identity
python-dotenv
requests